refactor(services/postgresql): split service into standalone crate
