+++ core/Cargo.toml
@@ -147,7 +147,7 @@ services-opfs = ["opendal-core/services-opfs"]
 services-oss = ["opendal-core/services-oss"]
 services-pcloud = ["opendal-core/services-pcloud"]
 services-persy = ["opendal-core/services-persy"]
-services-postgresql = ["opendal-core/services-postgresql"]
+services-postgresql = ["dep:opendal-service-postgresql"]
 services-redb = ["opendal-core/services-redb"]
 services-redis = ["opendal-core/services-redis"]
 services-redis-native-tls = ["opendal-core/services-redis-native-tls"]
@@ -210,6 +210,7 @@ opendal-service-ghac = { path = "services/ghac", version = "0.55.0", optional = 
 opendal-service-hdfs-native = { path = "services/hdfs-native", version = "0.55.0", optional = true, default-features = false }
 opendal-service-moka = { path = "services/moka", version = "0.55.0", optional = true, default-features = false }
 opendal-service-mysql = { path = "services/mysql", version = "0.55.0", optional = true, default-features = false }
+opendal-service-postgresql = { path = "services/postgresql", version = "0.55.0", optional = true, default-features = false }
 opendal-service-s3 = { path = "services/s3", version = "0.55.0", optional = true, default-features = false }
 
 [dev-dependencies]
