+++ core/src/lib.rs
@@ -43,6 +43,8 @@ pub mod services {
     pub use opendal_service_moka::*;
     #[cfg(feature = "services-mysql")]
     pub use opendal_service_mysql::*;
+    #[cfg(feature = "services-postgresql")]
+    pub use opendal_service_postgresql::*;
     #[cfg(feature = "services-s3")]
     pub use opendal_service_s3::*;
 }
